ocr_server/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pytesseract==0.3.13
Pillow==11.2.1
opencv-python==4.11.0.86
numpy==2.2.5
